refactor(tests): split into Unit and E2E suites; group Redis tests
- Move every test under tests/Cache/E2E so e2e and unit suites can be
  selected independently.
- Group RedisTest, RedisClusterTest under tests/Cache/E2E/Redis/.
- Update phpunit.xml to expose 'unit' and 'e2e' testsuites; the
  default test command no longer needs to enumerate files.
