+++ tests/Cache/E2E/MemoryTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Utopia\Tests;
+namespace Utopia\Tests\E2E;
 
 use Utopia\Cache\Adapter\Memory;
 use Utopia\Cache\Cache;
